# docs/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "version": "1.0.0",
    "title": "Contacts App",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "description": "A RESTful API for managing contacts, including authentication and user management."
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    },
    {
      "url": "https://nodejs-hw-mongodb-et1c.onrender.com"
    }
  ],
...
  "paths": {
    "/contacts": {
      "get": {
        "tags": [
          "Contacts"
        ],
        "summary": "Get list of contacts",
        "operationId": "getContacts",
        "description": "Retrieve a list of contacts using this endpoint.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer"
            },
            "example": 1
          },
          {
            "in": "query",
            "name": "perPage",
            "schema": {
              "type": "integer"
            },
            "example": 10
          },
          {
            "in": "query",
            "name": "sortBy",
            "schema": {
              "type": "string",
              "example": "name"
            },
            "description": "All contact fields can be mentioned. Default value - _id"
          },
          {
            "in": "query",
            "name": "sortOrder",
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ],
              "example": "asc"
            }
          },
          {
            "in": "query",
            "name": "contactType",
            "schema": {
              "type": "string",
              "enum": [
                "work",
                "home",
                "personal"
              ]
            },
            "description": "Filter contacts by type",
            "example": "home"
          },
          {
            "in": "query",
            "name": "isFavourite",
            "schema": {
              "type": "boolean"
            },
            "description": "Filter contacts by favourite status",
            "example": false
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully found contacts!",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully found contacts!"
                    },
                    "data": {
                      "type": "object",
                      "required": [
                        "data",
                        "page",
                        "perPage",
                        "totalItems",
                        "totalPages",
                        "hasNextPage",
                        "hasPrevPage"
                      ],
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/contacts"
                          }
                        },
                        "page": {
                          "type": "integer",
                          "example": 2
                        },
                        "perPage": {
                          "type": "integer",
                          "example": 5
                        },
                        "totalItems": {
                          "type": "integer",
                          "example": 20
                        },
                        "totalPages": {
                          "type": "integer",
                          "example": 4
                        },
                        "hasNextPage": {
                          "type": "boolean",
                          "example": false
                        },
                        "hasPrevPage": {
                          "type": "boolean",
                          "example": true
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 500
                    },
                    "message": {
                      "type": "string",
                      "example": "Internal server error"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Contacts"
        ],
        "summary": "Create contact",
        "operationId": "createContact",
        "description": "Create a contact with payload",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "phoneNumber",
                  "contactType"
                ],
                "properties": {
                  "name": {
                    "description": "Contact's name",
                    "type": "string",
                    "example": "Andrei Dumka"
                  },
                  "phoneNumber": {
                    "description": "Contact's phone number",
                    "type": "string",
                    "example": "++38091231230"
                  },
                  "email": {
                    "description": "Contact's email",
                    "type": "string",
                    "example": "andrei23415@mail.com"
                  },
                  "isFavourite": {
                    "description": "Is favorite contact",
                    "type": "boolean",
                    "example": true
                  },
                  "contactType": {
                    "description": "What type of contact does it belong to",
                    "type": "string",
                    "enum": [
                      "work",
                      "home",
                      "personal"
                    ],
                    "example": "home"
                  },
                  "userId": {
                    "description": "Contact's user id",
                    "type": "string",
                    "example": "683c67dfcff3a8e31ad7a061"
                  },
                  "photo": {
                    "description": "Contact's photo",
                    "type": "string",
                    "example": "https://res.cloudinary.com/dvgcgldeo/image/upload/v1749834846/photos/cykj7q14gasiokqcxqsv.jpg"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successfully created a contact",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 201,
                      "description": "Successfully created a contact!"
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully created a contact!",
                      "description": "Successfully created a contact!"
                    },
                    "data": {
                      "type": "object",
                      "$ref": "#/components/schemas/contacts"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
    "/contacts/{id}": {
      "get": {
        "tags": [
          "Contacts"
        ],
        "summary": "Get contact by id",
        "operationId": "getContact",
        "description": "Retrieves contact details by the specified contact ID.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "contactId",
            "required": true,
            "schema": {
              "type": "string",
              "example": "65caf65a087f10c88b598384"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully found contact by id",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully found contact with id 65caf65a087f10c88b598384!"
                    },
                    "data": {
                      "$ref": "#/components/schemas/contacts"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/404contactsId"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      },
      "put": {
        "tags": [
          "Contacts"
        ],
        "summary": "Replace contact by id",
        "operationId": "replaceContact",
        "description": "Replace the contact with the specified ID. This operation updates all fields and removes/reset missing ones.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "contactId",
            "required": true,
            "schema": {
              "type": "string",
              "example": "65caf65a087f10c88b598395"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "Contact's name",
                    "example": "Andrei Dumka"
                  },
                  "phoneNumber": {
                    "type": "string",
                    "description": "Contact's phone number",
                    "example": "+38091231230"
                  },
                  "email": {
                    "type": "string",
                    "format": "email",
                    "description": "Contact's email",
                    "example": "andrei23415@mail.com"
                  },
                  "isFavourite": {
                    "type": "boolean",
                    "description": "Is favorite contact",
                    "example": true
                  },
                  "contactType": {
                    "type": "string",
                    "enum": [
                      "work",
                      "home",
                      "personal"
                    ],
                    "description": "What type of contact does it belong to",
                    "example": "home"
                  },
                  "photo": {
                    "type": "string",
                    "description": "Contact's photo URL",
                    "example": "https://res.cloudinary.com/dvgcgldeo/image/upload/v1749834846/photos/cykj7q14gasiokqcxqsv.jpg"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Contact successfully replaced.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Contact successfully replaced with id 65caf65a087f10c88b598395!"
                    },
                    "data": {
                      "$ref": "#/components/schemas/contacts"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "404": {
            "$ref": "#/components/responses/404contactsId"
          }
        }
      },
      "patch": {
        "tags": [
          "Contacts"
        ],
        "summary": "Update contact by id",
        "operationId": "updateContact",
        "description": "Update the contact with the specified ID.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "contactId",
            "required": true,
            "schema": {
              "type": "string",
              "example": "65caf65a087f10c88b598395"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "Contact's name",
                    "example": "Andrei Dumka"
                  },
                  "phoneNumber": {
                    "type": "string",
                    "description": "Contact's phone number",
                    "example": "+3805574574577"
                  },
                  "email": {
                    "type": "string",
                    "description": "Contact's email",
                    "example": "andrei23415@mail.com"
                  },
                  "isFavourite": {
                    "type": "boolean",
                    "description": "Is favorite contact",
                    "example": true
                  },
                  "contactType": {
                    "type": "string",
                    "enum": [
                      "work",
                      "home",
                      "personal"
                    ],
                    "description": "What type of contact does it belong to",
                    "example": "home"
                  },
                  "photo": {
                    "type": "string",
                    "description": "Contact's photo URL",
                    "example": "https://res.cloudinary.com/dvgcgldeo/image/upload/v1749834846/photos/cykj7q14gasiokqcxqsv.jpg"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Contact successfully updated.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Contact successfully updated with id 65caf65a087f10c88b598384!"
                    },
                    "data": {
                      "$ref": "#/components/schemas/contacts"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "404": {
            "$ref": "#/components/responses/404contactsId"
          }
        }
      },
      "delete": {
        "tags": [
          "Contacts"
        ],
        "summary": "Delete contact by id",
        "operationId": "deleteContact",
        "description": "Delete the contact with the specified ID.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "contactId",
            "required": true,
            "schema": {
              "type": "string",
              "example": "68446e184c0a69111c1c5fb8"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successfully deleted contact — no content returned"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "404": {
            "$ref": "#/components/responses/404contactsId"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "contacts": {
        "type": "object",
        "required": [
          "name",
          "email",
          "phoneNumber",
          "contactType",
          "userId"
        ],
        "properties": {
          "_id": {
            "description": "Contact's id",
            "type": "string",
            "example": "68446e184c0a69111c1c5fb8"
          },
          "name": {
            "description": "Contact's name",
            "type": "string",
            "example": "Andrei Dumka"
          },
          "phoneNumber": {
            "description": "Contact's phone number",
            "type": "string",
            "example": "+38091231230"
          },
          "email": {
            "description": "Contact's email",
            "type": "string",
            "format": "email",
            "example": "andrei23415@mail.com"
          },
          "isFavourite": {
            "description": "Is favorite contact",
            "type": "boolean",
            "example": true
          },
          "contactType": {
            "description": "What type of contact does it belong to",
            "type": "string",
            "enum": [
              "work",
              "home",
              "personal"
            ],
            "example": "home"
          },
          "userId": {
            "description": "Contact's user id",
            "type": "string",
            "format": "uuid",
            "example": "683c67dfcff3a8e31ad7a061"
          },
          "photo": {
            "description": "Contact's photo",
            "type": "string",
            "example": "https://res.cloudinary.com/dvgcgldeo/image/upload/v1749834846/photos/cykj7q14gasiokqcxqsv.jpg"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "responses": {
      "400": {
        "description": "Bad request, missing or invalid parameters.",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 400,
                  "description": "HTTP status code indicating a bad request."
                },
                "message": {
                  "type": "string",
                  "example": "Invalid email or password format",
                  "description": "Detailed error message."
                }
              }
            }
          }
        }
      },
      "401": {
        "description": "Unauthorized error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message",
                "data"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 401
                },
                "message": {
                  "type": "string",
                  "example": "Unauthorized Error"
                },
                "data": {
                  "type": "object",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Access token expired"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "500": {
        "description": "Internal server error.",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 500,
                  "description": "HTTP status code indicating an internal server error."
                },
                "message": {
                  "type": "string",
                  "example": "Internal server error.",
                  "description": "A generic message indicating that the server encountered an unexpected error."
                },
                "errorCode": {
                  "type": "string",
                  "example": "SERVER_ERROR",
                  "description": "A specific error code to help identify the nature of the internal issue."
                },
                "method": {
                  "type": "string",
                  "example": "POST",
                  "description": "The HTTP method used when the error occurred."
                }
              }
            }
          }
        }
      },
      "404contactsId": {
        "description": "Contact not found",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message",
                "data"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 404,
                  "description": "HTTP status code indicating that the contact was not found."
                },
                "message": {
                  "type": "string",
                  "example": "Contact with contactId not found",
                  "description": "Contact with contactId not found"
                },
                "data": {
                  "type": "object",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Contact with contactId not found",
                      "description": "Contact with contactId not found"
                    }
                  }
                }
              }
            }
          }
        }
      },
...
    }
  }
}